+++ src/db-luna/MojDbLevelQuery.cpp
@@ -84,9 +84,9 @@ MojErr MojDbLevelQuery::getById(const MojObject& id, MojDbStorageItem*& itemOut,
         MojErrCheck(err);
         if (!foundOut) {
             char s[1024];
-            int size = primaryKey.size();
+            size_t size = primaryKey.size();
             (void) MojByteArrayToHex(primaryKey.data(), size, s);
-            MojLogInfo(MojDbLevelEngine::s_log, _T("bdbq_byId_warnindex: KeySize: %d; %s ;id: %s \n"),
+            MojLogInfo(MojDbLevelEngine::s_log, _T("bdbq_byId_warnindex: KeySize: %zu; %s ;id: %s \n"),
                                  size, s, primaryKey.data()+1);
 
             //MojErrThrow(MojErrDbInconsistentIndex);
